backend/requirements.txt
openai
python-dotenv
pdf2image
Pillow
flask
flask-cors
anthropic
werkzeug
